+++ hackerman/hackerman/algorithms.py
@@ -1,6 +1,0 @@
-def merge_list(list1, list2):
-	res = list1
-	for i in list2:
-		if not i in res:
-			res.append(i)
-	return res