+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.6.3
+	- HMC encryption now returnes bytes instead of str
+	- Added reverse TCP payload
+	- Added reverse TCP handler
+	- Added uid and safe_uid function to hackerman.utils
+
 # 0.6.2
 	- Bug fixes in hackerman.transport.tcp
 	- Added CHANGELOG.md to be used for update info